requirements.txt
discord.py==2.3.2
python-dotenv
pillow
aiohttp